% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{\%<+\%}
\alias{\%<+\%}
\title{\%<+\%}
\usage{
p \%<+\% data
}
\arguments{
\item{p}{ggplot2 object, such as ggtree or ggsc graphic object.}

\item{data}{data.frame, which must contains a column of \code{node},
or the first column of taxa labels, when \code{p} is a \code{ggtree} object.
Or it must contains columns of \code{.BarcodeID}, when \code{p} is a \code{ggsc}
object and \code{p$data} does not contain a column of \code{features}, if it
contains, the \code{data} must also contains a column of \code{features}.}
}
\value{
ggplot object with annotation data added
}
\description{
This operator attaches annotation data to a ggtree or ggsc graphic object
}
