% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqRaw.R
\docType{data}
\name{eqRaw}
\alias{eqRaw}
\title{A dataset containing historical earthquake record around the world from -70 to 2017 AD.}
\format{A data frame with 6099 data points, and 8 variables:
\describe{
\item{YEAR}{when the earthquake occur in year}
\item{MONTH}{when the earthquake occur in month}
\item{DAY}{when the earthquake occur in day}
\item{EQ_MAG_MS}{the surface-wave magnitude of the earthquake}
\item{COUNTRY}{the country where the earthquake occurred}
\item{LOCATION_NAME}{the detail location where the earthquake occurred}
\item{LATITUDE}{the latitude of the location}
\item{LONGITUDE}{the longitude of the location}
}}
\source{
https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\usage{
eqRaw
}
\description{
A dataset containing historical earthquake record around the world from -70 to 2017 AD.
}
\keyword{datasets}
