% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_set.R
\name{facet_set}
\alias{facet_set}
\title{facet_set}
\usage{
facet_set(label, side = "t", angle = NULL)
}
\arguments{
\item{label}{a character or a named vector to label the plot}

\item{side}{to label the plot at which side, either 't' (top) or 'r' (right)}

\item{angle}{angle of the facet label. Default is 0 for side='t' and -90 for side='r'.}
}
\value{
a ggplot with facet label
}
\description{
add a facet label to a ggplot or change facet label of a ggplot
}
