% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_image.R
\name{geom_image}
\alias{geom_image}
\title{(This function is not ready to use)}
\usage{
geom_image(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
Given a data set with the lat, long, group information to draw a US map and
change of party preference of each region, it draw an arrows for you.
}
\examples{

}
