% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-scatter-rect.R
\name{geom_scatter_rect}
\alias{geom_scatter_rect}
\title{geom_scatter_rect}
\usage{
geom_scatter_rect(
  mapping = NULL,
  data = NULL,
  asp = 0.6,
  width = 0.8,
  height = NULL,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping, default is NULL}

\item{data}{input data, default is NULL}

\item{asp}{aspect ration of rectangle box (height vs width), only works for height is missing}

\item{width}{width of the rectangles, default is 0.8}

\item{height}{height of the rectangles}

\item{...}{additional parameters passed to 'geom_rect'}
}
\description{
draw rectangle boxes as scatter points
}
\author{
Guangchuang Yu
}
