% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-volpoint.R
\name{geom_volpoint}
\alias{geom_volpoint}
\title{geom_volpoint}
\usage{
geom_volpoint(
  mapping = NULL,
  data = NULL,
  log2FC_cutoff = 2,
  p_cutoff = 1e-05,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{input data set}

\item{log2FC_cutoff}{cutoff values for log2FC}

\item{p_cutoff}{cutoff values p-value or adjusted p-value}

\item{...}{additional paramters passed to the layer}
}
\value{
a ggplot
}
\description{
layer of scatter points for volcano plot to visualize differential genes
}
