% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_plot_data}
\alias{get_plot_data}
\title{get_plot_data}
\usage{
get_plot_data(plot, var = NULL, layer = NULL)
}
\arguments{
\item{plot}{a 'gg' plot object}

\item{var}{variables to be extracted}

\item{layer}{specific layer to extract the data}
}
\value{
a data frame of selected variables
}
\description{
extract data from a 'gg' plot
}
\author{
Guangchuang Yu
}
