% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ggbreak2ggplot}
\alias{ggbreak2ggplot}
\title{ggbreak2ggplot}
\usage{
ggbreak2ggplot(plot)
}
\arguments{
\item{plot}{a ggbreak object}
}
\value{
a ggplot object
}
\description{
convert a ggbreak object to a ggplot object
}
\author{
Guangchuang Yu
}
