% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfun-package.R
\docType{package}
\name{ggfun-package}
\alias{ggfun}
\alias{ggfun-package}
\title{ggfun: Miscellaneous Functions for 'ggplot2'}
\description{
Useful functions and utilities for 'ggplot' object (e.g., geometric layers, themes, and utilities to edit the object).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/ggfun}
  \item Report bugs at \url{https://github.com/YuLab-SMU/ggfun/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
}

}
\keyword{internal}
