% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{yrange}
\alias{yrange}
\alias{xrange}
\alias{ggrange}
\title{plot range of a ggplot object}
\usage{
yrange(gg, type = "limit", region = "panel")

xrange(gg, type = "limit", region = "panel")

ggrange(gg, var, type = "limit", region = "panel")
}
\arguments{
\item{gg}{a ggplot object}

\item{type}{one of 'limit' or 'range', if 'region == "plot"',
to extract plot limit or plot data range}

\item{region}{one of 'panel' or 'plot' to indicate extracting range
based on the plot panel (scale expand will be counted) or
plot data (scale expand will not be counted)}

\item{var}{either 'x' or 'y'}
}
\value{
range of selected axis
}
\description{
extract x or y ranges of a ggplot
}
\author{
Guangchuang Yu
}
