% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.ggbreak}
\alias{is.ggbreak}
\title{is.ggbreak}
\usage{
is.ggbreak(plot)
}
\arguments{
\item{plot}{a plot obejct}
}
\value{
logical value
}
\description{
check whether a plot is a ggbreak object (including 'ggbreak', 'ggwrap' and 'ggcut' that defined in the 'ggbreak' package)
}
\author{
Guangchuang Yu
}
