% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybox.R
\name{keybox}
\alias{keybox}
\title{keybox}
\usage{
keybox(p, grob = "roundrect", gp = NULL)
}
\arguments{
\item{p}{a ggplot object}

\item{grob}{one of 'rect' or 'roundrect'}

\item{gp}{graphic parameter}
}
\value{
grob object
}
\description{
draw border for each of the ggplot legends
}
\author{
Guangchuang Yu
}
