% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-identify.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{identify}
\alias{gpar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{graphics}{\code{\link[graphics]{identify}}}

  \item{grid}{\code{\link[grid]{gpar}}}
}}

