% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_point_legend_shape}
\alias{set_point_legend_shape}
\title{set_point_legend_shape}
\usage{
set_point_legend_shape(plot)
}
\arguments{
\item{plot}{a 'gg' plot object}
}
\value{
an updated plot
}
\description{
override point legend set by 'aes(shape = I(shape))'
}
\author{
Guangchuang Yu
}
