% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_arrowmap.R
\name{stat_arrowmap}
\alias{stat_arrowmap}
\title{(This function is not ready to use)To Draw Arrow Map of US Election}
\usage{
stat_arrowmap(mapping = NULL, data = NULL, geom = "curve",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
Given a data set with the lat, long, group information to draw a US map and
change of party preference of each region, it draw an arrows for you.
}
\examples{

}
