% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ars.R
\name{stat_ars}
\alias{stat_ars}
\title{To Draw a Archimedean Spiral}
\usage{
stat_ars(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
To Draw a Archimedean Spiral
}
\examples{

ggplot() +
  stat_ars(aes(a = 10, b = 20), col = "blue")
}
