% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rl.R
\name{stat_rl}
\alias{stat_rl}
\title{To Draw a regression line for each level for categorical variable}
\usage{
stat_rl(mapping = NULL, data = NULL, geom = "line",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
To Draw a regression line for each level for categorical variable
}
\examples{
indexf <- read.table("index.txt", header = T),

ggplot(indexf, aes(x = length, y = width, id = sex)) +
geom_point() +
stat_rl(aes(x = length, y =  width, id = sex, colour = sex))
}
