% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_star.R
\name{stat_star}
\alias{stat_star}
\title{To Draw a Line from the central to the farest point}
\usage{
stat_star(mapping = NULL, data = NULL, geom = "line",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
Given a points set in data.frame, extends ggplot2 functionality to
draw a line from the central to the farest point
}
\examples{
madedata <- data.frame(x = rnorm(100, mean = 0, sd = 1),
y = rnorm(100, mean = 0, sd = 1))
ggplot(madedata, aes(x = x, y = y)) +
geom_point() +
stat_star(color = "red")


}
