% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_noxaxis}
\alias{theme_noxaxis}
\alias{theme_noyaxis}
\alias{theme_noaxis}
\title{theme_noxaxis}
\usage{
theme_noxaxis(color = "black", ...)

theme_noyaxis(color = "black", ...)

theme_noaxis(...)
}
\arguments{
\item{color}{color of y-axis}

\item{...}{additional parameters that passed to theme()}
}
\value{
ggplot2 theme
}
\description{
A theme that only show y-axis
}
\author{
Guangchuang Yu
}
