% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_fp}
\alias{theme_fp}
\title{theme_fp}
\usage{
theme_fp(x, i)
}
\arguments{
\item{x}{ggplot object to provide theme format}

\item{i}{the element of a theme provided by \code{x}}
}
\value{
theme element
}
\description{
theme format painter
}
\details{
It applies theme element (i) from a ggplot (x) to another ggplot object
}
\author{
Guangchuang Yu and Shuangbin Xu
}
