% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_no_margin}
\alias{theme_no_margin}
\title{theme_no_margin}
\usage{
theme_no_margin(...)
}
\arguments{
\item{...}{additional parameters that passed to theme()}
}
\value{
ggplot2 theme
}
\description{
A theme that has no margin
}
\author{
Guangchuang Yu
}
