% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_stamp}
\alias{theme_stamp}
\title{the theme of blind-like alias of theme_blinds}
\usage{
theme_stamp(colour = c("white", "grey"), axis = "y", ...)
}
\arguments{
\item{colour}{the colour of rectangular, default is c('white', 'grey60').}

\item{axis}{character which grid of axis will be filled, default is 'y'.}

\item{...}{additional parameters that passed to \code{theme} function.}
}
\description{
the theme of blind-like alias of theme_blinds
}
