% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-volpoint.R
\name{volplot}
\alias{volplot}
\title{volplot}
\usage{
volplot(data, mapping, log2FC_cutoff = 2, p_cutoff = 1e-05, ...)
}
\arguments{
\item{data}{input data set}

\item{mapping}{aesthetic mapping}

\item{log2FC_cutoff}{cutoff values for log2FC}

\item{p_cutoff}{cutoff values p-value or adjusted p-value}

\item{...}{additional paramters passed to the 'geom_volpoint' layer}
}
\value{
a ggplot
}
\description{
volcano plot
}
