% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-facet.R
\name{add_facet}
\alias{add_facet}
\title{add_facet}
\usage{
add_facet(plot, label)
}
\arguments{
\item{plot}{a ggplot object}

\item{label}{a string to label the plot}
}
\value{
a ggplot with facet label
}
\description{
add a facet label to a ggplot
}
\details{
add a facet label to a ggplot which only contains 1 panel
}
\author{
Guangchuang Yu
}
