% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqRaw.R
\docType{data}
\name{eqRaw}
\alias{eqRaw}
\title{A dataset containing historical earthquark record around the world from -70 to 2017 AD. Source: https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
\format{A data frame with 6099 data points, and 8 variables:
\describe{
\item{YEAR}{when the earthquark occur in year}
\item{MONTH}{when the earthquark occur in month}
\item{DAY}{when the earthquark occur in day}
\item{EQ_MAG_MS}{the surface-wave magnitude of the earthquake}
\item{COUNTRY}{the country where the earthquark occurred}
\item{LOCATION_NAME}{the detail location where the earthquark occurred}
\item{LATITUDE}{the latitude of the location}
\item{LONGITUDE}{the longitude of the location}

}}
\usage{
eqRaw
}
\description{
A dataset containing historical earthquark record around the world from -70 to 2017 AD. Source: https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\keyword{datasets}
