% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-legend.r
\name{get_legend}
\alias{get_legend}
\title{get_legend}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{a plot produce by grid, lattice, ggplot2 or composite plot}
}
\value{
a 'gtable' object of the legend
}
\description{
extract legend from a plot
}
\author{
Guangchuang Yu
}
